problems with builds when older version of cmake (2.6.x) is present
The cmake 2.6 does not add by default the libraries such as the math
library and the gfortran library (i.e. it does not add any standard
libraries). I have changed this behavior in the following way: if "the
version of cmake is < 2.8 and the compiler is gcc, then we add to the
link flags "-lm" and "-lgfortran".

The changes are in FASP_init.cmake and in test/CMakeLists.txt and
                                          tutorial/CMakeLists.txt

--Ludmil